OR billy2pupa - moved or metadata and created new init
